Uncouple handling of test protocol from the test tree
